fix: use numeric precision

+++ fuel-fetcher/db.sql
@@ -2,7 +2,7 @@ create table price (
     state text not null,
     station int not null,
     fuel text not null,
-    price float,
+    price numeric,
     changed_at int not null,
     checked_at int not null,
     primary key (state, station, fuel)
